# docs/swagger.yaml
basePath: /
definitions:
  entity.UserDTO:
    properties:
      login:
        maxLength: 30
        minLength: 6
        type: string
      password:
        minLength: 8
        type: string
    required:
    - login
    - password
    type: object
host: localhost:3000
info:
  contact:
    email: vadimocloc@gmail.com
    name: Kozenkov Vadim
    url: https://github.com/VadimOcLock
  description: API накопительной системы лояльности.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: GopherMart API
  version: v0.0.1
paths:
  /api/user/login:
    post:
      consumes:
      - application/json
      description: Аутентификация пользователя по логину и паролю.
      parameters:
      - description: Данные пользователя для входа
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.UserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно аутентифицирован.
          schema:
            type: string
        "400":
          description: Неверный формат запроса.
          schema:
            type: string
        "401":
          description: Неверная пара логин/пароль.
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера.
          schema:
            type: string
      summary: Аутентификация пользователя
      tags:
      - auth
  /api/user/register:
    post:
      consumes:
      - application/json
      description: Регистрация пользователя по логину и паролю. Логины должны быть
        уникальными. После успешной регистрации происходит автоматическая аутентификация.
      parameters:
      - description: Данные пользователя для регистрации
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.UserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно зарегистрирован и аутентифицирован.
          schema:
            type: string
        "400":
          description: Неверный формат запроса.
          schema:
            type: string
        "409":
          description: Логин уже занят.
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера.
          schema:
            type: string
      summary: Регистрация нового пользователя
      tags:
      - auth
swagger: "2.0"